+++ src/main/java/com/aaax/usecase/client/package-info.java
@@ -1,0 +1,2 @@
+/** Use cases for client. */
+package com.aaax.usecase.client;