+++ src/main/java/com/aaax/usecase/device/package-info.java
@@ -1,0 +1,2 @@
+/** Use cases for device. */
+package com.aaax.usecase.device;