+++ src/main/java/com/aaax/usecase/passkey/package-info.java
@@ -1,0 +1,2 @@
+/** Use cases for passkey. */
+package com.aaax.usecase.passkey;