+++ src/main/java/com/aaax/usecase/session/package-info.java
@@ -1,0 +1,2 @@
+/** Use cases for session. */
+package com.aaax.usecase.session;